requirements.txt
asgiref==3.2.10
Django==3.1
django-debug-toolbar==2.2
django-environ==0.4.5
django-model-utils==4.0.0
django-simple-history==2.11.0
Pillow==2.2.1
psycopg2==2.8.5
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
